+++ hephaistos-data/lua/Camera.lua
@@ -7,20 +7,13 @@ centered as Zagreus appears to exit the main play area.
 
 We simply scale down the soft clamp depending on the viewport scale factor,
 which allows the camera to behave identically to the original 16:9 viewport
-after being extended.
+after being extended. Also, when clamp weight is set to 0.0 as in Asphodel, we
+manually force it to 0.001 as otherwise it defaults to 1.0 and does weird things
+at higher resolutions. Funnily enough, the 0.001 kludge is already used by in
+the original Lua (`LeaveDeathAreaRoomPresentation` from `RoomPresentation.lua`).
 ]]
 
-local function rescaleSoftClamp(args)
-  args.SoftClamp = args.SoftClamp and args.SoftClamp * (1 / Hephaistos.ScaleFactorX) or nil
-end
-
-Hephaistos.RegisterFilterHook("SetCameraClamp", rescaleSoftClamp)
-
-local function hasClamp(args)
-  return args.SoftClamp
-end
-
-Hephaistos.SetCameraClamp[StartRoomPresentation] = hasClamp
-Hephaistos.SetCameraClamp[RestoreUnlockRoomExitsPresentation] = hasClamp
-Hephaistos.SetCameraClamp[StartDeathLoopPresentation] = hasClamp
-Hephaistos.SetCameraClamp[StartDeathLoopFromBoatPresentation] = hasClamp
+Hephaistos.RegisterPreHook("SetCameraClamp", function(args)
+  args.SoftClamp = args.SoftClamp == 0.0 and 0.001 or args.SoftClamp
+  args.SoftClamp = args.SoftClamp and args.SoftClamp * (1.0 / Hephaistos.ScaleFactorX) or args.SoftClamp
+end)