+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix camera clamping effect bugging out in instances where original clamp weight was `0.0`.
+
 ## [v1.3.1] - 2021-09-11
 
 ### Fixed